add check err for put and post methods

+++ src/backend/controllers/AccountController.js
@@ -76,48 +76,61 @@ class AccountController {
         if (req.get("Authorization")) {
             var userRole = jwt.verify(req.get("Authorization"), process.env.TOKEN_KEY).userRole
             if (userRole == 0 || userRole == 1 || userRole == 2) {
-                var accountData = req.body.accountData
-                Actors.findOne({username: accountData.username})
-                    .then(actor => {
-                        if (actor) {
-                            res.status(409).json('username existed')
-                        } else {
-                            if (userRole == 0) {
-                                const account = new Actors(accountData)
-                                account.save()
-                                    .then(() => {
-                                        Bases.updateOne({_id: accountData.workAt}, {$set: {managerID: account._id}})   
-                                            .then(() => {
-                                                res.status(200).json('Create Account Successfully')
-                                            })
-                                            .catch(err => {
-                                                res.status(500).json("Data invalid")
-                                            })
-                                    })
-                                    .catch(err => {
-                                        res.status(500).json("Data invalid")
-                                    })
+                if (req.body.accountData) {
+                    var accountData = req.body.accountData
+                    try {
+                        Actors.findOne({username: accountData.username})
+                        .then(actor => {
+                            if (actor) {
+                                res.status(409).json('username existed')
                             } else {
-                                var userBaseID = jwt.verify(req.get("Authorization"), process.env.TOKEN_KEY).workAt
-                                accountData.role = userRole == 1 ? 3 : 4
-                                accountData.workAt = userBaseID
-                                const account = new Actors(accountData)
-                                account.save()
-                                    .then(() => {
-                                        Bases.updateOne({_id: accountData.workAt}, {$push: {StaffIDs: account._id}})
-                                            .then(() => {
-                                                res.status(200).json('Create Account Successfully')
-                                            })
-                                            .catch(err => {
-                                                res.status(500).json("Data invalid")
-                                            })
-                                    })
-                                    .catch(err => {
-                                        res.status(500).json("Data invalid")
-                                    })
+                                if (userRole == 0) {
+                                    const account = new Actors(accountData)
+                                    account.save()
+                                        .then(() => {
+                                            Bases.updateOne({_id: accountData.workAt}, {$set: {managerID: account._id}})   
+                                                .then(() => {
+                                                    res.status(200).json('Create Account Successfully')
+                                                })
+                                                .catch(err => {
+                                                    console.log(err)
+                                                    res.status(400).json("Data invalid")
+                                                })
+                                        })
+                                        .catch(err => {
+                                            console.log(err)
+                                            res.status(400).json("Data invalid")
+                                        })
+                                } else {
+                                    var userBaseID = jwt.verify(req.get("Authorization"), process.env.TOKEN_KEY).workAt
+                                    accountData.role = userRole == 1 ? 3 : 4
+                                    accountData.workAt = userBaseID
+                                    const account = new Actors(accountData)
+                                    account.save()
+                                        .then(() => {
+                                            Bases.updateOne({_id: accountData.workAt}, {$push: {StaffIDs: account._id}})
+                                                .then(() => {
+                                                    res.status(200).json('Create Account Successfully')
+                                                })
+                                                .catch(err => {
+                                                    console.log(err)
+                                                    res.status(400).json("Data invalid")
+                                                })
+                                        })
+                                        .catch(err => {
+                                            console.log(err)
+                                            res.status(400).json("Data invalid")
+                                        })
+                                }
                             }
-                        }
-                    })
+                        })
+                    } catch (err) {
+                        console.log(err)
+                        res.status(400).json("Bad Request")
+                    }
+                } else {
+                    res.status(400).json("Account Data Empty")
+                }
             } else {
                 res.status(403).json('Permission Denied')
             }
@@ -131,12 +144,16 @@ class AccountController {
         if (req.get("Authorization")) {
             var userRole = jwt.verify(req.get("Authorization"), process.env.TOKEN_KEY).userRole
             if (userRole == 0 || userRole == 1 || userRole == 2) {
-                Actors.findById(req.query.id)
+                if (req.query.id) {
+                    Actors.findById(req.query.id)
                     .then(actor => {
                         res.status(200).json({
                             account: mongooseToObject(actor)
                         })
                     })
+                } else {
+                    res.status(404).json("Must specify account id")
+                }
             } else {
                 res.status(403).json('Permission Denied')
             }
@@ -150,36 +167,44 @@ class AccountController {
         if (req.get("Authorization")) {
             var userRole = jwt.verify(req.get("Authorization"), process.env.TOKEN_KEY).userRole
             if (userRole == 0 || userRole == 1 || userRole == 2) {
-                if (userRole == 0) {
-                    Actors.updateOne({_id: req.query.id}, req.body.accountData)
-                        .then(() => {
-                            Bases.updateOne({managerID: req.query.id}, {$unset: {managerID: ""}})
-                                .then(() => {
-                                    Bases.updateOne({_id: req.body.accountData.workAt}, {$set: {managerID: req.query.id}})
-                                        .then(() => {
-                                            res.status(200).json('Updated account successfully')
-                                        })
-                                        .catch(err => {
-                                            res.status(500).json("Data invalid")
-                                        })
-                                })
-                                .catch(err => {
-                                    res.status(500).json("Data invalid")
-                                })
-                        })
-                        .catch(err => {
-                            res.status(500).json("Data invalid")
-                        })
+                if (req.query.id) {
+                    if (userRole == 0) {
+                        Actors.updateOne({_id: req.query.id}, req.body.accountData)
+                            .then(() => {
+                                Bases.updateOne({managerID: req.query.id}, {$unset: {managerID: ""}})
+                                    .then(() => {
+                                        Bases.updateOne({_id: req.body.accountData.workAt}, {$set: {managerID: req.query.id}})
+                                            .then(() => {
+                                                res.status(200).json('Updated account successfully')
+                                            })
+                                            .catch(err => {
+                                                console.log(err)
+                                                res.status(400).json("Data invalid")
+                                            })
+                                    })
+                                    .catch(err => {
+                                        console.log(err)
+                                        res.status(400).json("Data invalid")
+                                    })
+                            })
+                            .catch(err => {
+                                console.log(err)
+                                res.status(400).json("Data invalid")
+                            })
+                    } else {
+                        Actors.updateOne(
+                            {_id: req.query.id}, 
+                            {$set: {name: req.body.accountData.name, username: req.body.accountData.username, password: req.body.accountData.password}})
+                            .then(() => {
+                                res.status(200).json('Updated account successfully')
+                            })
+                            .catch(err => {
+                                console.log(err)
+                                res.status(400).json("Data invalid")
+                            })
+                    }
                 } else {
-                    Actors.updateOne(
-                        {_id: req.query.id}, 
-                        {$set: {name: req.body.accountData.name, username: req.body.accountData.username, password: req.body.accountData.password}})
-                        .then(() => {
-                            res.status(200).json('Updated account successfully')
-                        })
-                        .catch(err => {
-                            res.status(500).json("Data invalid")
-                        })
+                    res.status(404).json("Must specify account id")
                 }
             } else {
                 res.status(403).json('Permission Denied')
@@ -194,30 +219,37 @@ class AccountController {
         if (req.get("Authorization")) {
             var userRole = jwt.verify(req.get("Authorization"), process.env.TOKEN_KEY).userRole
             if (userRole == 0 || userRole == 1 || userRole == 2) {
-                Actors.deleteMany({_id: {$in: req.body.accountIDs}})
-                    .then(() => {
-                        if (userRole == 0) {
-                            Bases.updateMany({managerID: {$in: req.body.accountIDs}}, {$unset: {managerID: ""}})
+                if (! req.body.accountIDs) {
+                    res.status(400).json("Must specify account id")
+                } else {
+                    Actors.deleteMany({_id: {$in: req.body.accountIDs}})
+                        .then(() => {
+                            if (userRole == 0) {
+                                Bases.updateMany({managerID: {$in: req.body.accountIDs}}, {$unset: {managerID: ""}})
+                                    .then(() => {
+                                        res.status(200).json({message: 'Delete Accounts Successfully'})
+                                    })
+                                    .catch(err => {
+                                        console.log(err)
+                                        res.status(400).json("Data invalid")
+                                    })
+                            } else {
+                                var userBaseID = jwt.verify(req.get("Authorization"), process.env.TOKEN_KEY).workAt
+                                Bases.updateOne({_id: userBaseID}, {$pull: {StaffIDs: {$in: req.body.accountIDs}}})
                                 .then(() => {
-                                    res.status(200).json({message: 'Delete Accounts Successfully'})
+                                    res.status(200).json({message: 'Delete Account Successfully'})
                                 })
                                 .catch(err => {
-                                    res.status(500).json("Data invalid")
+                                    console.log(err)
+                                    res.status(400).json("Data invalid")
                                 })
-                        } else {
-                            var userBaseID = jwt.verify(req.get("Authorization"), process.env.TOKEN_KEY).workAt
-                            Bases.updateOne({_id: userBaseID}, {$pull: {StaffIDs: {$in: req.body.accountIDs}}})
-                            .then(() => {
-                                res.status(200).json({message: 'Delete Account Successfully'})
-                            })
-                            .catch(err => {
-                                res.status(500).json("Data invalid")
-                            })
-                        }
-                    })
-                    .catch(err => {
-                        res.status(500).json("Data invalid")
-                    })
+                            }
+                        })
+                        .catch(err => {
+                            console.log(err)
+                            res.status(400).json("Data invalid")
+                        })
+                }
             } else {
                 res.status(403).json({message: 'Permission Denied'})
             }
